+++ frontend/public/changelog.md
@@ -10,6 +10,7 @@
 - Menambahkan autokomplet untuk pencarian
 - Mengubah pencarian glosarium menjadi mencari istilah Indonesia dan asing
 - Memfokuskan otomatis pada kotak cari saat membuka halaman
+- Menyesuaikan laman detail kamus
 
 ## 2026-02-14 Sabtu